Move modules to main file
Packaging functions in seprate files was causing
runtime error (Vosk engine was not being located).
Moving the modules to main.py resolved this issue.
