+++ main.py
@@ -1,5 +1,6 @@
 from vosk import Model, KaldiRecognizer
 import pyaudio
+import pyttsx3
 
 def model():
 
@@ -18,5 +19,18 @@ def model():
             break
 
         if recognizer.AcceptWaveform(data):
-            print(recognizer.Result())
             return recognizer.Result()
+
+
+def sayText(text):
+
+    engine = pyttsx3.init()
+    engine.setProperty('rate', 150)
+    engine.setProperty('volume',1.0)
+
+    voices = engine.getProperty('voices')
+    engine.setProperty('voice', voices[0].id)
+    engine.say(text)
+    engine.runAndWait()
+    engine.stop()
+    return text
