+++ vosk-model-en-us-0.22-lgraph/conf/mfcc.conf
