+++ vosk-model-en-us-0.22-lgraph/graph/phones.txt
