+++ vosk-model-en-us-0.22-lgraph/graph/phones/word_boundary.int
