+++ vosk-model-en-us-0.22-lgraph/ivector/final.dubm
