+++ vosk-model-en-us-0.22-lgraph/ivector/global_cmvn.stats
