+++ vosk-model-en-us-0.22-lgraph/ivector/online_cmvn.conf
